requirements.txt
pandas==2.1.1
numpy==1.26.0
pandas-ta==0.3.14b0
scipy==1.11.3
ccxt==4.0.98
python-telegram-bot==20.5
datetime==5.1
matplotlib==3.7.1